dev_requirements.txt
black==22.6.0 

python-dotenv
supervisely==6.68.26
imutils
pyzbar==0.1.9